refactor: reorganize API and types modules
