+++ src/types/rankings.ts
@@ -1,0 +1,56 @@
+import type { Team, User } from './user';
+
+export interface TopUsersResponse {
+  ranking: UserRanking[];
+  cursor: {
+    page: number;
+  };
+  total: number;
+}
+
+export interface UserRanking {
+  user: User;
+  ranked_score?: number;
+  pp?: number;
+}
+
+export interface CountryResponse {
+  ranking: CountryRanking[];
+  cursor: {
+    page: number;
+  };
+  total: number;
+}
+
+export interface CountryRanking {
+  code: string;
+  name: string;
+  active_users: number;
+  play_count: number;
+  ranked_score: number;
+  performance: number;
+}
+
+export type RankingType = 'performance' | 'score';
+export type TabType = 'users' | 'countries' | 'teams';
+
+export interface TeamRankingsResponse {
+  ranking: TeamRanking[];
+  cursor?: {
+    page: number;
+  };
+  total: number;
+}
+
+export interface TeamRanking {
+  team_id: number;
+  ruleset_id: number;
+  play_count: number;
+  ranked_score: number;
+  performance: number;
+}
+
+export interface TeamDetailResponse {
+  team: Team;
+  members: User[];
+}
