+++ src/types/ui.ts
@@ -1,0 +1,8 @@
+import type { ComponentType } from 'react';
+
+export interface NavItem {
+  path: string;
+  title: string;
+  icon?: ComponentType<{ size?: number; className?: string }>;
+  requireAuth?: boolean;
+}
