+++ src/utils/api/auth.ts
@@ -1,0 +1,67 @@
+import axios from 'axios';
+
+import { API_BASE_URL } from './client';
+
+export const authAPI = {
+  login: async (username: string, password: string, clientId: number, clientSecret: string) => {
+    console.log('Login attempt with:', { username, clientId });
+
+    const formData = new FormData();
+    formData.append('grant_type', 'password');
+    formData.append('client_id', clientId.toString());
+    formData.append('client_secret', clientSecret);
+    formData.append('username', username);
+    formData.append('password', password);
+    formData.append('scope', '*');
+
+    try {
+      const response = await axios.post(`${API_BASE_URL}/oauth/token`, formData, {
+        headers: {
+          'Content-Type': 'application/x-www-form-urlencoded',
+        },
+      });
+      return response.data;
+    } catch (error) {
+      const err = error as { response?: { data?: unknown }; message?: string };
+      console.error('Login error:', err.response?.data || err.message);
+      throw err;
+    }
+  },
+
+  register: async (username: string, email: string, password: string) => {
+    console.log('Register attempt with:', { username, email });
+
+    const formData = new URLSearchParams();
+    formData.append('user[username]', username);
+    formData.append('user[user_email]', email);
+    formData.append('user[password]', password);
+
+    try {
+      const response = await axios.post(`${API_BASE_URL}/users`, formData, {
+        headers: {
+          'Content-Type': 'application/x-www-form-urlencoded',
+        },
+      });
+      return response.data;
+    } catch (error) {
+      const err = error as { response?: { data?: unknown }; message?: string };
+      console.error('Register error:', err.response?.data || err.message);
+      throw err;
+    }
+  },
+
+  refreshToken: async (refreshToken: string, clientId: number, clientSecret: string) => {
+    const formData = new FormData();
+    formData.append('grant_type', 'refresh_token');
+    formData.append('client_id', clientId.toString());
+    formData.append('client_secret', clientSecret);
+    formData.append('refresh_token', refreshToken);
+
+    const response = await axios.post(`${API_BASE_URL}/oauth/token`, formData, {
+      headers: {
+        'Content-Type': 'application/x-www-form-urlencoded',
+      },
+    });
+    return response.data;
+  },
+};
