+++ src/utils/api/config.ts
@@ -1,0 +1,6 @@
+export const CLIENT_CONFIG = {
+  osu_client_id: 6,
+  osu_client_secret: '454532f3dba952663c8917ad15204d501ec2f28e41b3bce4b276b5d4a2a25823918f4711bb87a7fe',
+  web_client_id: 6,
+  web_client_secret: '454532f3dba952663c8917ad15204d501ec2f28e41b3bce4b276b5d4a2a25823918f4711bb87a7fe',
+};
